added an attempt at refactored generation logic for the READMEs
