+++ ebook-lint/cmd/epub/original.txt
@@ -1,0 +1,26 @@
+goos: linux
+goarch: amd64
+pkg: github.com/pjkaufman/go-go-gadgets/ebook-lint/cmd/epub
+cpu: AMD Ryzen 5 3500U with Radeon Vega Mobile Gfx  
+BenchmarkLintEpub-8   	       8	 140546692 ns/op
+BenchmarkLintEpub-8   	       9	 170948917 ns/op
+BenchmarkLintEpub-8   	       7	 174550400 ns/op
+BenchmarkLintEpub-8   	       6	 180642700 ns/op
+BenchmarkLintEpub-8   	       7	 183575685 ns/op
+BenchmarkLintEpub-8   	       7	 157349098 ns/op
+BenchmarkLintEpub-8   	       7	 156036569 ns/op
+BenchmarkLintEpub-8   	       7	 174257889 ns/op
+BenchmarkLintEpub-8   	       6	 174703443 ns/op
+BenchmarkLintEpub-8   	       6	 177452597 ns/op
+BenchmarkLintEpub-8   	       7	 186700484 ns/op
+BenchmarkLintEpub-8   	       7	 159547975 ns/op
+BenchmarkLintEpub-8   	       6	 180701655 ns/op
+BenchmarkLintEpub-8   	       6	 179066349 ns/op
+BenchmarkLintEpub-8   	       7	 174100705 ns/op
+BenchmarkLintEpub-8   	       6	 193817334 ns/op
+BenchmarkLintEpub-8   	       6	 190055322 ns/op
+BenchmarkLintEpub-8   	       6	 179252830 ns/op
+BenchmarkLintEpub-8   	       7	 177209681 ns/op
+BenchmarkLintEpub-8   	       7	 180461239 ns/op
+PASS
+ok  	github.com/pjkaufman/go-go-gadgets/ebook-lint/cmd/epub	32.936s
