+++ ebook-lint/cmd/epub/zip-to-zip.txt
@@ -1,0 +1,26 @@
+goos: linux
+goarch: amd64
+pkg: github.com/pjkaufman/go-go-gadgets/ebook-lint/cmd/epub
+cpu: AMD Ryzen 5 3500U with Radeon Vega Mobile Gfx  
+BenchmarkLintEpub-8   	       9	 150809896 ns/op
+BenchmarkLintEpub-8   	       7	 142989529 ns/op
+BenchmarkLintEpub-8   	       7	 174344400 ns/op
+BenchmarkLintEpub-8   	       8	 176714653 ns/op
+BenchmarkLintEpub-8   	       7	 153124364 ns/op
+BenchmarkLintEpub-8   	       6	 188455688 ns/op
+BenchmarkLintEpub-8   	       7	 153823553 ns/op
+BenchmarkLintEpub-8   	       6	 182980756 ns/op
+BenchmarkLintEpub-8   	       7	 179235346 ns/op
+BenchmarkLintEpub-8   	       7	 190940757 ns/op
+BenchmarkLintEpub-8   	       6	 168658193 ns/op
+BenchmarkLintEpub-8   	       7	 176506470 ns/op
+BenchmarkLintEpub-8   	       6	 173317566 ns/op
+BenchmarkLintEpub-8   	       6	 170153783 ns/op
+BenchmarkLintEpub-8   	       6	 187808469 ns/op
+BenchmarkLintEpub-8   	       7	 161630883 ns/op
+BenchmarkLintEpub-8   	       6	 176347434 ns/op
+BenchmarkLintEpub-8   	       7	 172040527 ns/op
+BenchmarkLintEpub-8   	       7	 155416536 ns/op
+BenchmarkLintEpub-8   	       6	 181028895 ns/op
+PASS
+ok  	github.com/pjkaufman/go-go-gadgets/ebook-lint/cmd/epub	35.663s
